requirements.txt
fastapi 
uvicorn 
transformers 
faiss-cpu 
pydantic
torch
python-dotenv
